requirements.txt
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==45.0.3
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
opencv-python==4.11.0.86
optionaldict==0.1.2
packaging==25.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pysmb==1.2.11
python-dotenv==1.1.0
requests==2.32.3
selenium==3.141.0
simplejson==3.20.1
SQLAlchemy==2.0.41
synology-drive-api==1.0.15
tqdm==4.67.1
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
Werkzeug==3.1.3
